// v3/wrap/src/__tests__/e2e/helpers/types.ts
export type UInt8 = number;
export type UInt16 = number;
export type UInt32 = number;
export type Int = number;
export type Int8 = number;
export type Int16 = number;
export type Int32 = number;
export type Bytes = Uint8Array;
export type BigInt = string;
export type Json = string;
export type String = string;
export type Boolean = boolean;

export interface SwapArgs {
  tokenIn: Token;
  tokenOut: Token;
  amount: BigInt;
  tradeType: TradeType;
}

export interface Currency {
  decimals: UInt8;
  symbol?: String | null;
  name?: String | null;
}

export interface Token {
  chainId: ChainId;
  address: String;
  currency: Currency;
}

export interface Price {
  baseToken: Token;
  quoteToken: Token;
  denominator: BigInt;
  numerator: BigInt;
  price: string;
}

export interface Fraction {
  denominator: BigInt;
  numerator: BigInt;
  quotient: string;
}

export interface Route {
  pools: Array<Pool>;
  path: Array<Token>;
  input: Token;
  output: Token;
  midPrice: Price;
}

export interface Tick {
  index: Int32;
  liquidityGross: BigInt;
  liquidityNet: BigInt;
}

export interface Pool {
  token0: Token;
  token1: Token;
  fee: FeeAmount;
  sqrtRatioX96: BigInt;
  liquidity: BigInt;
  tickCurrent: Int32;
  tickDataProvider: Tick[];
  token0Price: Price;
  token1Price: Price;
}

export interface Trade {
  swaps: Array<TradeSwap>;
  tradeType: TradeType;
  inputAmount: TokenAmount;
  outputAmount: TokenAmount;
}

export interface TokenAmount {
  token: Token;
  amount: BigInt;
}

export interface TradeSwap {
  route: Route;
  inputAmount: TokenAmount;
  outputAmount: TokenAmount;
}
// ...
export interface MethodParameters {
  calldata: String;
  value: String;
}
// ...
export enum ChainIdEnum {
  MAINNET,
  ROPSTEN,
  RINKEBY,
  GOERLI,
  KOVAN,
}

export type ChainIdString =
  | "MAINNET"
  | "ROPSTEN"
  | "RINKEBY"
  | "GOERLI"
  | "KOVAN"

export type ChainId = ChainIdEnum | ChainIdString;

export enum FeeAmountEnum {
  LOWEST,
  LOW,
  MEDIUM,
  HIGH,
}

export type FeeAmountString =
  | "LOWEST"
  | "LOW"
  | "MEDIUM"
  | "HIGH"

export type FeeAmount = FeeAmountEnum | FeeAmountString;

export enum TradeTypeEnum {
  EXACT_INPUT,
  EXACT_OUTPUT,
}

export type TradeTypeString =
  | "EXACT_INPUT"
  | "EXACT_OUTPUT"

export type TradeType = TradeTypeEnum | TradeTypeString;
// ...
export interface Ethers_TxResponse {
  hash: String;
  to?: String | null;
  from: String;
  nonce: UInt32;
  gasLimit: BigInt;
  gasPrice?: BigInt | null;
  data: String;
  value: BigInt;
  chainId: UInt32;
  blockNumber?: BigInt | null;
  blockHash?: String | null;
  timestamp?: UInt32 | null;
  confirmations: UInt32;
  raw?: String | null;
  r?: String | null;
  s?: String | null;
  v?: UInt32 | null;
  type?: UInt32 | null;
  accessList?: Array<Ethers_Access> | null;
}

/* URI: "w3://ens/ethereum.web3api.eth" */
export interface Ethers_Access {
  address: String;
  storageKeys: Array<String>;
}

/* URI: "w3://ens/ethereum.web3api.eth" */
export interface Ethers_TxReceipt {
  to: String;
  from: String;
  contractAddress: String;
  transactionIndex: UInt32;
  root?: String | null;
  gasUsed: BigInt;
  logsBloom: String;
  transactionHash: String;
  logs: Array<Ethers_Log>;
  blockNumber: BigInt;
  blockHash: String;
  confirmations: UInt32;
  cumulativeGasUsed: BigInt;
  effectiveGasPrice: BigInt;
  byzantium: Boolean;
  type: UInt32;
  status?: UInt32 | null;
}

/* URI: "w3://ens/ethereum.web3api.eth" */
export interface Ethers_Log {
  blockNumber: BigInt;
  blockHash: String;
  transactionIndex: UInt32;
  removed: Boolean;
  address: String;
  data: String;
  topics: Array<String>;
  transactionHash: String;
  logIndex: UInt32;
}